audit fixes

+++ programs/charcoin/src/burn.rs
@@ -1,8 +1,8 @@
+use crate::ConfigAccount;
 use anchor_lang::prelude::*;
 use anchor_lang::solana_program::clock::Clock;
-use anchor_spl::token::{Mint, TokenAccount};
 use anchor_spl::token::{self, Burn, Token};
-use crate::ConfigAccount;
+use anchor_spl::token::{Mint, TokenAccount};
 
 #[derive(Accounts)]
 pub struct ExecuteBuyback<'info> {
@@ -10,15 +10,15 @@ pub struct ExecuteBuyback<'info> {
         mut,
         seeds=[b"config".as_ref()],
         bump
-      )]    
+      )]
     pub config_account: Account<'info, ConfigAccount>,
     #[account(mut)]
     pub mint: Account<'info, Mint>,
     #[account(mut)]
-    pub burn_wallet_ata: Account<'info,TokenAccount>,
+    pub burn_wallet_ata: Account<'info, TokenAccount>,
     #[account(
         mut,
-        constraint = config_account.config.death_wallet == burn_authority.key() // Ensure the signer is the admin
+        constraint = config_account.config.death_wallet == burn_authority.key() 
     )]
     pub burn_authority: Signer<'info>,
 
@@ -30,17 +30,10 @@ pub enum CustomError {
     #[msg("No tokens available for buyback.")]
     NoTokensToBuyback,
 }
-
-pub fn execute_buyback(
-    ctx: Context<ExecuteBuyback>
-) -> Result<()> {
-    
-
+// this function will be run in the backend inside a cron job
+pub fn execute_buyback(ctx: Context<ExecuteBuyback>) -> Result<()> {
     let tokens_to_buy = ctx.accounts.burn_wallet_ata.amount;
-    require!(
-        tokens_to_buy > 0,
-        CustomError::NoTokensToBuyback
-    );
+    require!(tokens_to_buy > 0, CustomError::NoTokensToBuyback);
     // Burn tokens from the burn_wallet.
     let burn_ctx = CpiContext::new(
         ctx.accounts.token_program.to_account_info(),
@@ -57,7 +50,7 @@ pub fn execute_buyback(
     tracker.total_burned += tokens_to_buy;
 
     // Cache the current timestamp.
-    let current_time = Clock::get()?.unix_timestamp;
+    let current_time = Clock::get()?.unix_timestamp as u64;
 
     // Emit an event logging the buyback and burn details.
     emit!(BuybackBurnEvent {
@@ -65,7 +58,7 @@ pub fn execute_buyback(
         new_total_burned: tracker.total_burned,
         timestamp: current_time,
     });
-   
+
     Ok(())
 }
 
@@ -73,5 +66,5 @@ pub fn execute_buyback(
 pub struct BuybackBurnEvent {
     pub tokens_bought: u64,
     pub new_total_burned: u64,
-    pub timestamp: i64,
+    pub timestamp: u64,
 }
